PilotSpeak.NET now targets .NET 6.0! - Removed a lot of boilerplate code that is no longer necessary - Statements from the ``Main()`` method have been moved to the top-level - No longer on PeachPie SDK preview! - Simplified project names
