+++ .github/workflows/dotnet.yml
@@ -13,7 +13,7 @@ jobs:
         runs-on: ${{ matrix.os }}
         strategy:
             matrix:
-                dotnet: ["3.1.x", "5.0.x"]
+                dotnet: ["6.0.x"]
                 os: [ubuntu-latest, windows-latest, macOS-latest]
 
         steps:
