+++ .vscode/tasks.json
@@ -1,4 +1,6 @@
 {
+  // See https://go.microsoft.com/fwlink/?LinkId=733558
+  // for the documentation about the tasks.json format
   "version": "2.0.0",
   "tasks": [
     {
@@ -7,7 +9,7 @@
       "type": "process",
       "args": [
         "build",
-        "${workspaceFolder}/src/Raccoony.PilotSpeak.Server/Raccoony.PilotSpeak.Server.csproj",
+        "${workspaceFolder}/src/PilotSpeak.Server/PilotSpeak.Server.csproj",
         "/property:GenerateFullPaths=true",
         "/consoleloggerparameters:NoSummary"
       ],
@@ -19,7 +21,7 @@
       "type": "process",
       "args": [
         "publish",
-        "${workspaceFolder}/src/Raccoony.PilotSpeak.Server/Raccoony.PilotSpeak.Server.csproj",
+        "${workspaceFolder}/src/PilotSpeak.Server/PilotSpeak.Server.csproj",
         "/property:GenerateFullPaths=true",
         "/consoleloggerparameters:NoSummary"
       ],
@@ -32,9 +34,8 @@
       "args": [
         "watch",
         "run",
-        "${workspaceFolder}/src/Raccoony.PilotSpeak.Server/Raccoony.PilotSpeak.Server.csproj",
-        "/property:GenerateFullPaths=true",
-        "/consoleloggerparameters:NoSummary"
+        "--project",
+        "${workspaceFolder}/src/PilotSpeak.Server/PilotSpeak.Server.csproj"
       ],
       "problemMatcher": "$msCompile"
     }
