+++ changelog.md
@@ -1,0 +1,22 @@
+# Change Log
+
+## 1.0.100
+
+PilotSpeak.NET now targets .NET 6.0!
+
+- Removed a lot of boilerplate code that is no longer necessary
+- Statements from the ``Main()`` method have been moved to the top-level
+- No longer on PeachPie SDK preview!
+- Simplified project names
+
+## 0.1.100
+
+- Added Docker support
+- Moved CSS and JS files to wwwroot directory
+- Use .NET SDK [versioning conventions](https://docs.microsoft.com/en-us/dotnet/core/versions/)
+- Replaced ``wordbank.php`` with ``wordbank.json``
+- Renamed top wordbank variable to getjson because that variable already exists further down
+
+### Note
+
+Technically speaking there was no ``0.1.x`` up until ``1.0.x``. These notes were retroactively added to the changelog for the sake of completeness.
