+++ src/PilotSpeak.Server/PilotSpeak.Server.csproj
@@ -2,19 +2,20 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-        <AssemblyName>pilotspeak</AssemblyName>
+        <TargetFramework>net6.0</TargetFramework>
+        <Version>0.1.200</Version>
         <UserSecretsId>283754b1-fd11-42e1-b958-1909cef67281</UserSecretsId>
         <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
+        <ImplicitUsings>enable</ImplicitUsings>
     </PropertyGroup>
 
     <ItemGroup>
         <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.10.9" />
-        <PackageReference Include="Peachpie.AspNetCore.Web" Version="1.0.0-preview1" />
+        <PackageReference Include="Peachpie.AspNetCore.Web" Version="1.0.17" />
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Raccoony.PilotSpeak\Raccoony.PilotSpeak.msbuildproj" />
+        <ProjectReference Include="..\PilotSpeak\PilotSpeak.msbuildproj" />
     </ItemGroup>
 
 </Project>
