+++ src/PilotSpeak.Server/Program.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore;
+using PilotSpeak.Server;
+
+var host = WebHost.CreateDefaultBuilder(args)
+    .UseStartup<Startup>()
+    .Build();
+
+host.Run();
