+++ src/PilotSpeak/PilotSpeak.msbuildproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Peachpie.NET.Sdk/1.0.17">
+    <PropertyGroup>
+      <OutputType>Library</OutputType>
+      <TargetFramework>net6.0</TargetFramework>
+    </PropertyGroup>
+    <ItemGroup>
+      <Compile Include="**/*.php;**/*.phar" />
+    </ItemGroup>
+    <ItemGroup>
+      <Content Include="**" Exclude="obj/**;bin/**;PublishProfiles/**;@(Compile);$(MsBuildThisFile)">
+        <PackagePath>contentFiles/any/netcoreapp3.0/wwwroot/</PackagePath>
+        <PackageCopyToOutput>true</PackageCopyToOutput>
+      </Content>
+    </ItemGroup>
+  </Project>
