+++ src/PilotSpeak/index.php
@@ -9,7 +9,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
     <script src="konami.js"></script>
-    <script src="https://kit.fontawesome.com/0528c2bae4.js" crossorigin="anonymous"></script>
+    <script src="https://kit.fontawesome.com/0cb6746fd4.js" crossorigin="anonymous"></script>
     <link href="https://fonts.googleapis.com/css?family=Comfortaa|Roboto&display=swap" rel="stylesheet">
     <link href="style.css" rel="stylesheet">
     <title>Pilot Speak Generator</title>
@@ -60,10 +60,10 @@
         $totalcomboformatted = number_format($totalcombo);
         $wordbank = count($descriptor) + count($time) + count($location_noun1) + count($buzzword1);
         $wordbankformatted = number_format($wordbank);?>
-      <p><small>So far, <mark><strong><?php echo $yes ?></strong></mark> phrases were generated using this app. Application made in love by <a href="http://t.me/drraccoon/">@DrRaccoon</a> for <a href="http://t.me/vulan/">@Vulan</a> and all my other aviation friends. Like what I make? <a href="https://ko-fi.com/drraccoon"><i class="fad fa-mug-hot"></i> Buy me a coffee</a>!<br>
+      <p><small>So far, <mark><strong><?php echo $yes ?></strong></mark> phrases were generated using this app. Application made in love by <a href="https://twitter.com/DrRaccoon">@DrRaccoon</a> for <a href="https://twitter.com/FlyinFox53">@FlyinFox53</a> and ported to .NET by <a href="https://twitter.com/tonyt1ns">@tonyt1ns</a>.
+      Like what I make? <a href="https://ko-fi.com/drraccoon"><i class="fa-solid fa-cup-togo"></i> Buy me a coffee</a>!<br>
         There are currenly <mark><? echo $wordbankformatted; ?></mark> words, making <mark><? echo $totalcomboformatted; ?></mark> total combinations. Want to contribute more phases and words? Join this <a href="https://t.me/PilotSpeak">telegram group</a>. Last updated: <mark><? echo "".date("F d Y.",filemtime("wordbank.json")); ?></mark></small></p>
         <center><p>Accidentally made in <i class="fab fa-php"></i>
-        <br /><small>Accidentally ported to <i class="fab fa-microsoft"></i> .NET by Tonytins</small></center></p>
      </div>
      <input type="text" value="The <?php echo $buzzword; ?> on <?php echo $location; ?> <?php echo $time; ?> was <?php echo $descriptor; ?>." id="textInput" style="opacity: 0;">
      <script>
@@ -80,7 +80,7 @@
 
      <script>
        konami.load();
-       var easter_egg = new Konami('http://your-special-easter-egg-website.com');
+       var easter_egg = new Konami('https://www.furaffinity.net/view/26114232/');
      </script>
   </body>
 </html>
