--- src/Raccoony.PilotSpeak.Server/Program.cs
@@ -1,19 +1,0 @@
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
-
-namespace Raccoony.PilotSpeak.Server
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            var host = WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
-                .Build();
-
-            host.Run();
-        }
-    }
-}
