--- src/Raccoony.PilotSpeak.Server/Startup.cs
@@ -1,39 +1,0 @@
-using System;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace Raccoony.PilotSpeak.Server
-{
-    class Startup
-    {
-        public void ConfigureServices(IServiceCollection services)
-        {
-            // Adds a default in-memory implementation of IDistributedCache.
-            services.AddDistributedMemoryCache();
-
-            services.AddSession(options =>
-            {
-                options.IdleTimeout = TimeSpan.FromMinutes(30);
-                options.Cookie.HttpOnly = true;
-            });
-
-            services.AddPhp(options =>
-            {
-
-            });
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-                app.UseDeveloperExceptionPage();
-
-            app.UseSession();
-            app.UsePhp();
-            app.UseDefaultFiles();
-            app.UseStaticFiles();
-        }
-    }
-}
