--- src/Raccoony.PilotSpeak/Raccoony.PilotSpeak.msbuildproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Peachpie.NET.Sdk/1.0.0-preview1">
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <AssemblyName>Raccoony.PilotSpeakGenerator</AssemblyName>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="**/*.php;**/*.phar" />
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="**" Exclude="obj/**;bin/**;PublishProfiles/**;@(Compile);$(MsBuildThisFile)">
-      <PackagePath>contentFiles/any/netcoreapp3.0/wwwroot/</PackagePath>
-      <PackageCopyToOutput>true</PackageCopyToOutput>
-      <CopyToOutputDirectory>true</CopyToOutputDirectory>
-      <CopyToPublishDirectory>true</CopyToPublishDirectory>
-    </Content>
-  </ItemGroup>
-</Project>
